+++ jamdict/jmdict.py
@@ -127,6 +127,9 @@ class KanjiForm(object):
     fields. Synonyms are not included; they may be indicated in the
     cross-reference field associated with the sense element.
     DTD <!ELEMENT k_ele (keb, ke_inf*, ke_pri*)>
+    text --- a kanji written form of an entry, string
+    info --- coded information field, a list of strings
+    pri --- relative priority of the entry, a list of strings
     '''
 
     def __init__(self, text=''):
@@ -200,7 +203,13 @@ class KanaForm(object):
     Where there are multiple reading elements, they will typically be
     alternative readings of the kanji element. In the absence of a
     kanji element, i.e. in the case of a word or phrase written
-    entirely in kana, these elements will define the entry.'''
+    entirely in kana, these elements will define the entry.
+    text --- a kana written form of an entry, string
+    nokanji --- True means this entry cannot be regarded as a true reading of the kanji, boolean
+    restr --- use to restrict the reading to a subset of the available kanji forms, list of string
+    info --- coded information field, a list of strings
+    pri --- relative priority of the entry, a list of strings
+    '''
 
     def __init__(self, text='', nokanji=False):
         '''this element content is restricted to kana and related
@@ -435,6 +444,30 @@ class Sense(object):
         return sd
 
 
+class Translation(Sense):
+    ''' The trans element will record the translational equivalent
+    of the Japanese name, plus other related information. (JMendict)
+    <!ELEMENT trans (name_type*, xref*, trans_det*)>'''
+    def __init__(self):
+        super().__init__()
+        self.name_type = []  # mapped to name_type*
+        self.xref = []  # mapped to xref
+        self.gloss = []  # mapped to trans_det
+
+    def name_type_human(self):
+        return [JMENDICT_TYPE_MAP[x] if x in JMENDICT_TYPE_MAP else x for x in self.name_type]
+
+    def text(self, compact=True):
+        tmp = [str(x) for x in self.gloss]
+        types = "/".join(self.name_type) if compact else "/".join(self.name_type_human())
+        return '{gloss} ({types})'.format(gloss='/'.join(tmp), types=types)
+
+    def to_json(self):
+        sd = super().to_json()
+        sd['name_type'] = self.name_type
+        return sd
+
+
 class SenseGloss(object):
     '''Within each sense will be one or more "glosses", i.e.
         target-language words or phrases which are equivalents to the
@@ -466,7 +499,8 @@ class SenseGloss(object):
 
     def __str__(self):
         tmp = [self.text]
-        if self.lang:
+        if self.lang and self.lang != 'eng':
+            # lang = eng is trivial
             tmp.append('(lang:%s)' % self.lang)
         if self.gend:
             tmp.append('(gend:%s)' % self.gend)
@@ -515,6 +549,23 @@ class LSource:
                 'lstype': self.lstype,
                 'wasei': self.wasei,
                 'text': self.text}
+
+
+JMENDICT_TYPES = (("surname", "family or surname"),
+                  ("place", "place name"),
+                  ("unclass", "unclassified name"),
+                  ("company", "company name"),
+                  ("product", "product name"),
+                  ("work", "work of art, literature, music, etc. name"),
+                  ("masc", "male given name or forename"),
+                  ("fem", "female given name or forename"),
+                  ("person", "full name of a particular person"),
+                  ("given", "given name or forename, gender not specified"),
+                  ("station", "railway station"),
+                  ("organization", "organization name"),
+                  ("ok", "old or irregular kana form"))
+JMENDICT_TYPE_MAP = dict(JMENDICT_TYPES)
+JMENDICT_TYPE_MAP_DECODE = {v: k for k, v in JMENDICT_TYPES}
 
 
 class Meta(object):
@@ -568,6 +619,9 @@ class JMDictXMLParser(object):
                 self.parse_info(child, entry)
             elif child.tag == 'sense':
                 self.parse_sense(child, entry)
+            elif child.tag == 'trans':
+                # JMendict support
+                self.parse_ne_translation(child, entry)
             else:
                 raise Exception("Invalid tag: %s" % child.tag)
         return entry
@@ -656,6 +710,24 @@ class JMDictXMLParser(object):
                 raise Exception("WARNING: invalid tag in bibinfo (child.tag = %s)" % child.tag)
         entry_info.bibinfo.append(bib)
         return bib
+
+    def parse_ne_translation(self, trans_tag, entry):
+        translation = Translation()
+        for child in trans_tag:
+            if child.tag == 'name_type':
+                _name_type = JMENDICT_TYPE_MAP_DECODE[child.text] if child.text in JMENDICT_TYPE_MAP_DECODE else child.text
+                translation.name_type.append(_name_type)
+            elif child.tag == 'trans_det':
+                # add sensegloss
+                lang = self.get_attrib(trans_tag, 'xml:lang', default_value='eng')
+                gloss = SenseGloss(lang=lang, gend='', text=child.text)
+                translation.gloss.append(gloss)
+            elif child.tag == 'xref':
+                translation.xref.append(child.text)
+            else:
+                raise Exception("Invalid tag: {} in JMendict/trans tag".format(child.tag))
+        entry.senses.append(translation)
+        return translation
 
     def parse_sense(self, sense_tag, entry):
         sense = Sense()