+++ CHANGES.md
@@ -1,5 +1,11 @@
+2020-05-31
+    - [Version 0.1a7]
+    - Added Japanese Proper Names Dictionary (JMnedict) support
+    - Included built-in KRADFILE/RADKFile support
+    - Improved command line tools (json, compact mode, etc.)
+
 2017-08-18
-        - Support for KanjiDic2 (XML/SQLite formats)
+    - Support KanjiDic2 (XML/SQLite formats)
         
 2016-11-09
-	- Release first demo to Github
+    - Release first version to Github